# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Lazy AI Coder - GitLab integration, LLM tools, and MCP server\n\nFeatures:\n- LLM chat and code analysis\n- GitLab integration (read files, review MRs)\n- MCP (Model Context Protocol) server\n- PlantUML diagram generation\n- Session management with memory",
        "title": "Lazy AI Coder API",
        "contact": {
            "name": "API Support",
            "email": "walter.fan@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/smart-prompt/auto-fill-fields": {
            "post": {
                "description": "Use LLM to automatically fill framework fields based on user input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Auto-fill framework fields with LLM",
                "parameters": [
                    {
                        "description": "Auto-fill Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartprompt.AutoFillFieldsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/smartprompt.AutoFillFieldsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/smart-prompt/frameworks": {
            "get": {
                "description": "Get list of all available prompt engineering frameworks (CRISPE, RISEN, CO-STAR, APE, CAR)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Get all prompt engineering frameworks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smartprompt.Framework"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a custom prompt engineering framework",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Create a new framework",
                "parameters": [
                    {
                        "description": "Framework data",
                        "name": "framework",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartprompt.Framework"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/smartprompt.Framework"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/smart-prompt/frameworks/{id}": {
            "get": {
                "description": "Get details of a specific prompt engineering framework",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Get framework by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Framework ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/smartprompt.Framework"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a prompt engineering framework by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Update an existing framework",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Framework ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated framework data",
                        "name": "framework",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartprompt.Framework"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/smartprompt.Framework"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a prompt engineering framework by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Delete a framework",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Framework ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/smart-prompt/generate": {
            "post": {
                "description": "Generate CAR format prompt with GitLab context analysis, code examples, and quality scoring",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Generate smart prompt with context detection",
                "parameters": [
                    {
                        "description": "Smart Prompt Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SmartPromptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SmartPromptResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/smart-prompt/generate-from-framework": {
            "post": {
                "description": "Generate a prompt using a specific framework and field values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Generate prompt from framework",
                "parameters": [
                    {
                        "description": "Generation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartprompt.GenerateFromFrameworkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/smart-prompt/quick-refine": {
            "post": {
                "description": "Get framework-specific quick suggestions without calling LLM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-prompt"
                ],
                "summary": "Get quick refinement suggestions",
                "parameters": [
                    {
                        "description": "Quick Refinement Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smartprompt.QuickRefineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "smartprompt.AutoFillFieldsRequest": {
            "type": "object",
            "required": [
                "framework_id",
                "user_input"
            ],
            "properties": {
                "framework_id": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/models.Settings"
                },
                "user_input": {
                    "type": "string"
                }
            }
        },
        "smartprompt.AutoFillFieldsResponse": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}